+++ origo.js
@@ -23,6 +23,7 @@ origo.controls.sharemap = require('./src/sharemap');
 origo.controls.legend = require('./src/legend');
 origo.controls.search = require('./src/search');
 origo.controls.editor = require('./src/editor');
+origo.controls.home = require('./src/home');
 
 origo.map.init = function(el, options) {
     var map = mapLoader(el, options);
