+++ src/viewer.js
@@ -66,7 +66,6 @@ function init(el, mapOptions) {
     settings.center = urlParams.center || mapOptions.center;
     settings.zoom = urlParams.zoom || mapOptions.zoom;
     settings.source = mapOptions.source;
-    settings.home = mapOptions.home;
     settings.groups = mapOptions.groups;
     settings.editLayer = mapOptions.editLayer;
     settings.styles = mapOptions.styles;
@@ -104,7 +103,6 @@ function init(el, mapOptions) {
         mapwindow.init();
     }
 
-    createHome(settings.home);
     loadMap();
 
     //Check size for attribution mode
@@ -908,20 +906,6 @@ function init(el, mapOptions) {
               map.removeOverlay(overlays[i]);
             }
         }
-    }
-    function createHome(home) {
-        var el = utils.createButton({
-            id: 'home-button',
-            iconCls: 'mdk-icon-fa-home',
-            src: 'css/svg/fa-icons.svg#fa-home',
-            tooltipText: 'Zooma till hela kartan'
-        });
-        $('#map').append(el);
-        $('#home-button').on('touchend click', function(e) {
-          map.getView().fit(home, map.getSize());
-          $('#home-button button').blur();
-          e.preventDefault();
-        });
     }
     function checkSize() {
         var small = map.getSize()[0] < 768;
